+++ app/build.gradle
@@ -50,6 +50,7 @@ dependencies {
     implementation "androidx.work:work-runtime-ktx:2.7.1"
     implementation 'com.google.code.gson:gson:2.8.8'
     implementation "androidx.multidex:multidex:2.0.1"
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
