+++ app/src/main/AndroidManifest.xml
@@ -2,11 +2,10 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
-    <uses-permission android:name="android.permission.INTERNET"/>
-    <uses-permission android:name="android.permission.FOREGROUND_SERVICE"/>
+    <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.FOREGROUND_SERVICE" />
     <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
-    <uses-permission android:name="android.permission.REQUEST_IGNORE_BATTERY_OPTIMIZATIONS"/>
-
+    <uses-permission android:name="android.permission.REQUEST_IGNORE_BATTERY_OPTIMIZATIONS" />
     <uses-permission android:name="android.permission.VIBRATE" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
@@ -16,52 +15,50 @@
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_main"
         android:label="@string/app_name"
+        android:requestLegacyExternalStorage="true"
         android:roundIcon="@mipmap/ic_main"
         android:supportsRtl="true"
         android:theme="@style/Theme.NightViewer"
-        android:requestLegacyExternalStorage="true"
         tools:targetApi="31">
-
-
-
-
         <activity
-            android:name=".FullscreenActivity"
-            android:configChanges="orientation|keyboardHidden|screenSize|smallestScreenSize"
-            android:exported="true"
-            android:launchMode="singleTop"
-            android:theme="@style/Theme.NightViewer.Fullscreen"
-            android:resizeableActivity="true">
+            android:name=".StartURLActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
         <activity
+            android:name=".FullscreenActivity"
+            android:configChanges="orientation|keyboardHidden|screenSize|smallestScreenSize"
+            android:exported="true"
+            android:launchMode="singleTop"
+            android:resizeableActivity="true"
+            android:theme="@style/Theme.NightViewer.Fullscreen">
+<!--            <intent-filter>-->
+<!--                <action android:name="android.intent.action.MAIN" />-->
+<!--                <category android:name="android.intent.category.LAUNCHER" />-->
+<!--            </intent-filter>-->
+        </activity>
+        <activity
             android:name=".FullscreenActivity1"
             android:configChanges="keyboardHidden"
             android:launchMode="singleTop"
-            android:theme="@style/Theme.NightViewer.Fullscreen"
-            android:minHeight="100dp"
             android:minWidth="100dp"
-            android:resizeableActivity="true">
-
-        </activity>
+            android:minHeight="100dp"
+            android:resizeableActivity="true"
+            android:theme="@style/Theme.NightViewer.Fullscreen"></activity>
         <activity
             android:name=".FullscreenActivity2"
             android:configChanges="keyboardHidden"
             android:launchMode="singleTop"
-            android:theme="@style/Theme.NightViewer.Fullscreen"
-            android:resizeableActivity="true">
-
-        </activity>
-
+            android:resizeableActivity="true"
+            android:theme="@style/Theme.NightViewer.Fullscreen"></activity>
         <activity
             android:name=".PreferencesActivity"
-            android:screenOrientation="portrait"
             android:exported="false"
-            android:label="@string/title_activity_preferences" />
-
+            android:label="@string/title_activity_preferences"
+            android:screenOrientation="portrait" />
     </application>
 
 </manifest>
