+++ app/src/main/java/com/nightscout/nightviewer/BGData.kt
@@ -2,7 +2,6 @@ package com.nightscout.nightviewer
 
 import android.content.Context
 import android.util.Log
-import android.widget.Toast
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import org.json.JSONObject
@@ -10,7 +9,7 @@ import java.net.URL
 import java.text.SimpleDateFormat
 
 class BGData(private val context: Context){
-    private val pref_urlText = prefs.getString("ns_url", "https://pkd7320591.my.nightscoutpro.com")
+    val pref_urlText = prefs.getString("ns_url", "defaultURL")
     //User_Prefs : 기본 Preferences (iob, cob, basal enable 등)
     //BG_db: 혈당 데이터베이스 (Preferences로 구현 sql로도 가능할듯)
     fun initializeBG_db(){ //최근 10개 데이터로 db initialize //delta,arrow정보는 제외
@@ -54,7 +53,7 @@ class BGData(private val context: Context){
                 }
             }
             else{
-                showErrorMessage(context, "ERROR")
+//                showErrorMessage(context, "ERROR")
             }
         }.start()
     }
@@ -93,7 +92,7 @@ class BGData(private val context: Context){
             null
         }
         if (url == null) {
-            Toast.makeText(context, "데이터를 불러올 수 없습니다. 확인을 누른 후 앱을 종료합니다.", Toast.LENGTH_SHORT).show()
+//            Toast.makeText(context, "데이터를 불러올 수 없습니다. 확인을 누른 후 앱을 종료합니다.", Toast.LENGTH_SHORT).show()
             return null
         }
         val result = URL(url.toString()).readText()
