+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,3 @@
-
 <resources>
     <string name="app_name">NightViewer</string>
     <string name="title_activity_preferences">Preferences</string>
